Create run-tests.yml
Note: In the example above, the workflow runs when we push to the session-1 branch or make a pull request to the main branch. 

There is one job called test which has four steps. The first two use the built-in checkout and setup-python GitHub actions. The third step installs the Python prerequisites using Python pip. That last step runs the tests. These last two steps run the same commands you would run on your computer if you wanted to run the tests manually.